+++ src/main/java/me/zort/commandlib/CommandLibProxy.java
@@ -1,6 +1,5 @@
 package me.zort.commandlib;
 
-import me.zort.commandlib.internal.CommandEntry;
 import net.md_5.bungee.api.ChatColor;
 import net.md_5.bungee.api.CommandSender;
 import net.md_5.bungee.api.ProxyServer;
